move custom type guards to types file
